allow for parameters to be passed to script
added parameter for the two file types to be passed.  If nothing passed
will take the default zipcode.csv and name_cc.csv files in same
directory as script.  Still need to add the # of trans and # of
transactions before a fraud occurs but plan on doing that tonight.
Script should still be in working order now.